pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["ciaran28 <ciaranh@microsoft.com>"]
readme = "README.md"
packages = [
    {include = "data", from = "src"},
    {include = "ai", from = "src"},
    {include = "bi", from = "src"}
]

[tool.poetry.scripts]
s_a = "src.py_main_sa_actions:run"
d_a = "src.ai.py_main_da:run_depression_analysis"
bi = "src.bi.sentiment_insights:run_bi"
eval = "src.llmops.evaluate:run_evaluation_metrics"
actions = "src.common.actions:run_retrieve_actions"



[tool.poetry.dependencies]
python = "^3.10"
ipdb = "^0.13.13"
numpy = "^2.2.0"
pandas = "^2.2.3"
deltalake = "^0.22.3"
semantic-kernel = "^1.17.0"
transformers = "^4.47.0"
torch = "^2.5.1"
tiktoken = "^0.8.0"
protobuf = "^5.29.1"
nltk = "^3.9.1"
wordcloud = "^1.9.4"
seaborn = "^0.13.2"
scikit-learn = "^1.6.0"
xgboost = "^2.1.3"
python-dotenv = "^1.0.1"
rich = "^13.9.4"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
